Reorganize the tests in IncrementalCompilationTests.swift
